fix: linting errors
